+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.3 (unreleased)
 ----------------
 
+- Introducting new scale direction "autodown". If no cropped scale is available
+  deliver direction="down" scale.
+  [petschki]
+
 - Purge proxy caches if needed on crop.
   [alecm]